// front-end/packages/shared/src/models/paymentStatus.ts
/**
 * Generated by Orval
 * Do not edit manually.
 * OpenAPI specification - RING-Bookstore
 */

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  PENDING: "PENDING",
  CANCELLED: "CANCELLED",
  UNDERPAID: "UNDERPAID",
  PAID: "PAID",
  EXPIRED: "EXPIRED",
  PROCESSING: "PROCESSING",
  FAILED: "FAILED",
  PENDING_REFUND: "PENDING_REFUND",
  REFUNDED: "REFUNDED",
} as const;